runtime: Limit spmc channel max receivers
Limited spmc channel max receivers to `u16` from `usize` as it is more than enough and to avoid runtime failures.

+++ component_integration_tests/rust_test_scenarios/src/tests/runtime/channel/spmc_broadcast.rs
@@ -19,7 +19,7 @@ use tracing::info;
 struct TestInput {
     data_to_send: Vec<u64>,
     receivers: Vec<String>,
-    max_receiver_count: usize,
+    max_receiver_count: u16,
 }
 
 impl TestInput {
@@ -116,7 +116,7 @@ impl Scenario for SPMCBroadcastSendReceive {
 struct OverflowTestInput {
     population_method: String,
     receiver_count: usize,
-    max_receiver_count: usize,
+    max_receiver_count: u16,
 }
 
 impl OverflowTestInput {
@@ -563,7 +563,7 @@ impl Scenario for SPMCBroadcastDropSender {
 struct HeavyTestInput {
     send_count: u64,
     receivers: Vec<String>,
-    max_receiver_count: usize,
+    max_receiver_count: u16,
 }
 
 impl HeavyTestInput {
